+++ src/js/mixin/togglable.js
@@ -24,6 +24,7 @@ export default {
         cls: Boolean,
         animation: 'list',
         duration: Number,
+        velocity: Number,
         origin: String,
         transition: String,
     },
@@ -32,8 +33,9 @@ export default {
         cls: false,
         animation: [false],
         duration: 200,
+        velocity: 0.2,
         origin: false,
-        transition: 'linear',
+        transition: 'ease',
         clsEnter: 'uk-togglabe-enter',
         clsLeave: 'uk-togglabe-leave',
 
@@ -44,6 +46,7 @@ export default {
             paddingBottom: '',
             marginTop: '',
             marginBottom: '',
+            boxShadow: '',
         },
 
         hideProps: {
@@ -53,6 +56,7 @@ export default {
             paddingBottom: 0,
             marginTop: 0,
             marginBottom: 0,
+            boxShadow: 'none',
         },
     },
 
@@ -147,7 +151,15 @@ export default {
     },
 };
 
-export function toggleHeight({ isToggled, duration, initProps, hideProps, transition, _toggle }) {
+export function toggleHeight({
+    isToggled,
+    duration,
+    velocity,
+    initProps,
+    hideProps,
+    transition,
+    _toggle,
+}) {
     return (el, show) => {
         const inProgress = Transition.inProgress(el);
         const inner = el.hasChildNodes()
@@ -168,6 +180,8 @@ export function toggleHeight({ isToggled, duration, initProps, hideProps, transi
         fastdom.flush();
 
         const endHeight = height(el) + (inProgress ? 0 : inner);
+        duration = velocity * el.offsetHeight + duration;
+
         height(el, currentHeight);
 
         return (